disable spring sequrity for this project. Add new features: transfer money, view all transactions, view accaount info, view balance.

+++ backend/src/main/java/com/alex/bankingservicesapp/models/BankUser.java
@@ -46,10 +46,10 @@ public class BankUser {
 
 
     @Column(name = "user_balance")
-    private long balance;
+    private Double balance;
 
 
-    public BankUser(String accountName, String firstName, String lastName, String email, String password, long balance) {
+    public BankUser(String accountName, String firstName, String lastName, String email, String password, Double balance) {
         this.accountName = accountName;
         this.firstName = firstName;
         this.lastName = lastName;
